refactor: refactored folders structure and added split contants
