+++ src/components/List/post-list.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import PostListItem from '../post-list-item';
+import PostListItem from '../ListItem';
 import './post-list.css'
 
 const PostList = ({posts, onDelete, onToggleImportant, onToggleLiked}) => {
