+++ src/components/SearchPanel/search-panel.css
