+++ src/components/SearchPanel/search-panel.js
