+++ src/components/SearchPanel/search-panel.scss
