+++ src/shared/posts.js
@@ -1,0 +1,50 @@
+
+import githubPic from '../assets/4.png';
+import flexboxMem from '../assets/1.png';
+import udemyMem from '../assets/3.png';
+import gotMem from '../assets/123.jpg';
+
+export const posts = [
+    {
+        label: 'Наколькі добрая ідэя шукаць працу праз мемы? 🤔',
+        important: false,
+        like: false,
+        id: 1
+    },
+    {
+        label: 'Адзін лайк і я пачну так рабіць! Пагналі!',
+        important: false,
+        like: false,
+        id: 2
+    },
+    {
+        label: <img className='pic' src={githubPic} alt="img" />,
+        important: false,
+        like: false,
+        id: 3
+    },
+    {
+        label: <img className='pic' src={flexboxMem} alt="img" />,
+        important: false,
+        like: false,
+        id: 4
+    },
+    {
+        label: 'Самы час ці то лайкнуць, ці то дэлітнуць 🧐',
+        important: false,
+        like: false,
+        id: 5
+    },
+    {
+        label: <img className='pic' src={udemyMem} alt="img" />,
+        important: false,
+        like: false,
+        id: 6
+    },
+    {
+        label: <img className='pic' src={gotMem} alt="img" />,
+        important: false,
+        like: false,
+        id: 7
+    },
+];
